+++ MultiCommTerminal/WindowObjs/TerminatorDataSelector.xaml.cs
@@ -77,7 +77,9 @@ namespace MultiCommTerminal.WindowObjs {
         private void btnDelete_Click(object sender, RoutedEventArgs e) {
             var item = this.listBoxTerminators.SelectedItem as IIndexItem<DefaultFileExtraInfo>;
             if (item != null) {
-                this.wrapper.DeleteTerminatorData(item, this.ReloadList, App.ShowMsg);
+                if (MsgBoxYesNo.ShowBoxDelete(this, item.Display) == MsgBoxYesNo.MsgBoxResult.Yes) {
+                    this.wrapper.DeleteTerminatorData(item, this.ReloadList, App.ShowMsg);
+                }
             }
         }
